refactor versioning
