+++ lib/scala/text-buffer/src/main/scala/org/enso/text/ContentBasedVersioning.scala
@@ -1,4 +1,4 @@
-package org.enso.languageserver.data
+package org.enso.text
 
 /** A content-based versioning calculator.
   */
