+++ lib/scala/text-buffer/src/main/scala/org/enso/text/Sha3_224VersionCalculator.scala
@@ -1,4 +1,4 @@
-package org.enso.languageserver.data
+package org.enso.text
 
 import org.bouncycastle.jcajce.provider.digest.SHA3
 import org.bouncycastle.util.encoders.Hex
